Find suitable update image
Loops through images in the update 'till it finds one at least 400x200 to use as the update image.

If none are found, it'll log it but take no further action - I believe we have a default image in place already.

Should prevent banners/emotes being picked up for the update image.